Item Genericizing and offhand shooting
The player now has different slots for items in a dictionary instead of specific vars for guns etc.
Picking up a gun is now generic for all items.
Removed PickupWeapon, as it basically did nothing and can just be replaced with the base Pickup.
Made control action names generic for items. "shoot" is now "use", etc.
Right mouse click or right trigger will now either use the item held in the right hand, or do an alt use command on the item in the left hand for zooming etc.