Exits feature complete for demo
You can now have multiple exits that go to the same room. As long as the destination room has matching exits that go back to the room you're currently in and are in the same relative location to each other, the exit you're going through will figure out which exit to spawn you at on the other side. Works for up to 4 players.