Upgrade Godot to 4.0.2
Upgraded Godot project to 4.0.2. Pogs all around.
Known issues:
Dialogue Manager addon was updated and basically everything related to dialogue needs to be redone.
Signals were changed and now heart icons don't animate properly.
Zombie pathing wasn't tested, but pathing was changed so I suspect work will have to be done on that.
There is an issue with the unselected tabs in the inventory rendering above the panel.